now project can be created by a mvn clean install command, and tests are setuped properly

+++ dto/src/main/java/org/hotiver/dto/seller/SellerProfileDto.java
@@ -4,6 +4,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -11,7 +13,7 @@ public class SellerProfileDto {
     private Long id;
     private String displayName;
     private String nickname;
-    private Double rating;
+    private BigDecimal rating;
     private String profileDescription;
 
 }
